Add files via upload
This is a simple calculator that is meant to aid math practice. It uses sympy to show the final answers that one should get when dealing with integrals or derivatives. The project also uses regex and kivy to make sympy easily usable even to those with no technical skills at all.